+++ src/index.ts
@@ -1,3 +1,11 @@
-import * as Signature from './signature';
+import { secureCompare, sign, validate } from './signature';
+import { generateFetcher } from './fetcher';
+import { translateParamsToQuery } from './helpers/paramsEncoder';
 
-export { Signature };
+export {
+  secureCompare,
+  sign,
+  validate,
+  generateFetcher,
+  translateParamsToQuery,
+};