+++ src/types/paramsTypes.ts
@@ -6,11 +6,9 @@ export type ParamsTypes =
       [key: string]: ParamsValueType | undefined | null;
     };
 
-// Type for parameters cleaned from undefined and null values
-export type ParamsHashNotNilType = { [key: string]: ParamsValueType };
-
 // Type for parameters values
 export type ParamsValueType =
+  | boolean
   | number
   | string
   | Array<ParamsValueType>