+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [1.0.0] - 2022-09-01
+
+## Added
+
+- `generateFetcher` function have been added.
+- Helpers functions have been exposed.
+
+## Changed
+
+- Signature container has been removed and all methods are flattened.
+
 ## [0.2.0] - 2022-08-31
 
 ### Added
@@ -42,7 +53,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Basic skeleton for the package
 
-[unreleased]: https://github.com/Billcorporate/m2m_keygen_ts/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/Billcorporate/m2m_keygen_ts/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/Billcorporate/m2m_keygen_ts/releases/tag/v1.0.0
 [0.2.0]: https://github.com/Billcorporate/m2m_keygen_ts/releases/tag/v0.2.0
 [0.1.3]: https://github.com/Billcorporate/m2m_keygen_ts/releases/tag/v0.1.3
 [0.1.2]: https://github.com/Billcorporate/m2m_keygen_ts/releases/tag/v0.1.2